c13 - building, starting porting poa specific cli stuff
