+++ bin/reth/src/commands/debug_cmd/build_block.rs
@@ -309,7 +309,7 @@ impl Command {
                 let executor_factory = EvmProcessorFactory::new(self.chain.clone(), evm_config);
                 let mut executor = executor_factory.with_state(blockchain_db.latest()?);
                 executor
-                    .execute_and_verify_receipt(&block_with_senders.clone().unseal(), U256::MAX)?;
+                    .execute_and_verify_receipt(&block_with_senders.clone().unseal(), U256::MAX, None)?;
                 let state = executor.take_output_state();
                 debug!(target: "reth::cli", ?state, "Executed block");
