+++ bin/reth/src/commands/poa/mod.rs
@@ -1,0 +1,397 @@
+//! Main node command
+//!
+//! Starts the client
+
+use clap::{value_parser, Parser};
+use reth_authority_consensus::AuthorityConsensus;
+use reth_auto_seal_consensus::AutoSealConsensus;
+use reth_beacon_consensus::BeaconConsensus;
+use reth_interfaces::consensus::Consensus;
+use reth_primitives::ChainSpec;
+use std::{net::SocketAddr, path::PathBuf, sync::Arc};
+
+pub mod cl_events;
+pub mod events;
+pub mod notifications;
+
+use crate::{
+    args::{
+        utils::{chain_help, genesis_value_parser, parse_socket_address, SUPPORTED_CHAINS},
+        DatabaseArgs, DebugArgs, DevArgs, NetworkArgs, PayloadBuilderArgs, PruningArgs,
+        RpcServerArgs, TxPoolArgs,
+    },
+    builder::{launch_from_config, NodeConfig},
+    cli::{db_type::DatabaseBuilder, ext::RethCliExt},
+    dirs::{DataDirPath, MaybePlatformPath},
+    runner::CliContext,
+};
+
+/// Start the node
+#[derive(Debug, Parser)]
+pub struct PoaNodeCommand<Ext: RethCliExt = ()> {
+    /// The path to the data dir for all reth files and subdirectories.
+    ///
+    /// Defaults to the OS-specific data directory:
+    ///
+    /// - Linux: `$XDG_DATA_HOME/reth/` or `$HOME/.local/share/reth/`
+    /// - Windows: `{FOLDERID_RoamingAppData}/reth/`
+    /// - macOS: `$HOME/Library/Application Support/reth/`
+    #[arg(long, value_name = "DATA_DIR", verbatim_doc_comment, default_value_t)]
+    pub datadir: MaybePlatformPath<DataDirPath>,
+
+    /// The path to the configuration file to use.
+    #[arg(long, value_name = "FILE", verbatim_doc_comment)]
+    pub config: Option<PathBuf>,
+
+    /// The chain this node is running.
+    ///
+    /// Possible values are either a built-in chain or the path to a chain specification file.
+    #[arg(
+        long,
+        value_name = "CHAIN_OR_PATH",
+        long_help = chain_help(),
+        default_value = SUPPORTED_CHAINS[0],
+        default_value_if("dev", "true", "dev"),
+        value_parser = genesis_value_parser,
+        required = false,
+    )]
+    pub chain: Arc<ChainSpec>,
+
+    /// Enable Prometheus metrics.
+    ///
+    /// The metrics will be served at the given interface and port.
+    #[arg(long, value_name = "SOCKET", value_parser = parse_socket_address, help_heading = "Metrics")]
+    pub metrics: Option<SocketAddr>,
+
+    /// Add a new instance of a node.
+    ///
+    /// Configures the ports of the node to avoid conflicts with the defaults.
+    /// This is useful for running multiple nodes on the same machine.
+    ///
+    /// Max number of instances is 200. It is chosen in a way so that it's not possible to have
+    /// port numbers that conflict with each other.
+    ///
+    /// Changes to the following port numbers:
+    /// - DISCOVERY_PORT: default + `instance` - 1
+    /// - AUTH_PORT: default + `instance` * 100 - 100
+    /// - HTTP_RPC_PORT: default - `instance` + 1
+    /// - WS_RPC_PORT: default + `instance` * 2 - 2
+    #[arg(long, value_name = "INSTANCE", global = true, default_value_t = 1, value_parser = value_parser!(u16).range(..=200))]
+    pub instance: u16,
+
+    /// Overrides the KZG trusted setup by reading from the supplied file.
+    #[arg(long, value_name = "PATH")]
+    pub trusted_setup_file: Option<PathBuf>,
+
+    /// All networking related arguments
+    #[clap(flatten)]
+    pub network: NetworkArgs,
+
+    /// All rpc related arguments
+    #[clap(flatten)]
+    pub rpc: RpcServerArgs,
+
+    /// All txpool related arguments with --txpool prefix
+    #[clap(flatten)]
+    pub txpool: TxPoolArgs,
+
+    /// All payload builder related arguments
+    #[clap(flatten)]
+    pub builder: PayloadBuilderArgs,
+
+    /// All debug related arguments with --debug prefix
+    #[clap(flatten)]
+    pub debug: DebugArgs,
+
+    /// All database related arguments
+    #[clap(flatten)]
+    pub db: DatabaseArgs,
+
+    /// All pruning related arguments
+    #[clap(flatten)]
+    pub pruning: PruningArgs,
+
+    /// Additional cli arguments
+    #[clap(flatten)]
+    #[clap(next_help_heading = "Extension")]
+    pub ext: Ext::Node,
+}
+
+impl<Ext: RethCliExt> PoaNodeCommand<Ext> {
+    /// Replaces the extension of the node command
+    pub fn with_ext<E: RethCliExt>(self, ext: E::Node) -> PoaNodeCommand<E> {
+        let Self {
+            datadir,
+            config,
+            chain,
+            metrics,
+            trusted_setup_file,
+            instance,
+            network,
+            rpc,
+            txpool,
+            builder,
+            debug,
+            db,
+            pruning,
+            #[cfg(feature = "optimism")]
+            rollup,
+            ..
+        } = self;
+        PoaNodeCommand {
+            datadir,
+            config,
+            chain,
+            metrics,
+            instance,
+            trusted_setup_file,
+            network,
+            rpc,
+            txpool,
+            builder,
+            debug,
+            db,
+            pruning,
+            #[cfg(feature = "optimism")]
+            rollup,
+            ext,
+        }
+    }
+
+    /// Execute `node` command
+    pub async fn execute(self, ctx: CliContext) -> eyre::Result<()> {
+        let Self {
+            datadir,
+            config,
+            chain,
+            metrics,
+            trusted_setup_file,
+            instance,
+            network,
+            rpc,
+            txpool,
+            builder,
+            debug,
+            db,
+            pruning,
+            #[cfg(feature = "optimism")]
+            rollup,
+            ext,
+        } = self;
+
+        // set up real database
+        let database = DatabaseBuilder::Real(datadir);
+
+        // set up node config
+        let node_config = NodeConfig {
+            database,
+            config,
+            chain,
+            metrics,
+            instance,
+            trusted_setup_file,
+            network,
+            rpc,
+            txpool,
+            builder,
+            debug,
+            db,
+            dev: DevArgs::default(),
+            pruning,
+            #[cfg(feature = "optimism")]
+            rollup,
+        };
+
+        let executor = ctx.task_executor;
+
+        // launch the node
+        let handle = launch_from_config::<Ext>(node_config, ext, executor).await?;
+
+        handle.wait_for_node_exit().await
+    }
+
+    /// Returns the [Consensus] instance to use.
+    pub fn consensus(&self) -> Arc<dyn Consensus> {
+        let poa_consensus = Arc::new(AuthorityConsensus::new(self.chain.clone()));
+
+        poa_consensus
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::args::utils::SUPPORTED_CHAINS;
+    use reth_discv4::DEFAULT_DISCOVERY_PORT;
+    use std::{
+        net::{IpAddr, Ipv4Addr},
+        path::Path,
+    };
+
+    #[test]
+    fn parse_help_node_command() {
+        let err = PoaNodeCommand::<()>::try_parse_from(["reth", "--help"]).unwrap_err();
+        assert_eq!(err.kind(), clap::error::ErrorKind::DisplayHelp);
+    }
+
+    #[test]
+    fn parse_common_node_command_chain_args() {
+        for chain in SUPPORTED_CHAINS {
+            let args: PoaNodeCommand = PoaNodeCommand::<()>::parse_from(["reth", "--chain", chain]);
+            assert_eq!(args.chain.chain, chain.parse::<reth_primitives::Chain>().unwrap());
+        }
+    }
+
+    #[test]
+    fn parse_discovery_addr() {
+        let cmd = PoaNodeCommand::<()>::try_parse_from(["reth", "--discovery.addr", "127.0.0.1"])
+            .unwrap();
+        assert_eq!(cmd.network.discovery.addr, Ipv4Addr::LOCALHOST);
+    }
+
+    #[test]
+    fn parse_addr() {
+        let cmd = PoaNodeCommand::<()>::try_parse_from([
+            "reth",
+            "--discovery.addr",
+            "127.0.0.1",
+            "--addr",
+            "127.0.0.1",
+        ])
+        .unwrap();
+        assert_eq!(cmd.network.discovery.addr, Ipv4Addr::LOCALHOST);
+        assert_eq!(cmd.network.addr, Ipv4Addr::LOCALHOST);
+    }
+
+    #[test]
+    fn parse_discovery_port() {
+        let cmd =
+            PoaNodeCommand::<()>::try_parse_from(["reth", "--discovery.port", "300"]).unwrap();
+        assert_eq!(cmd.network.discovery.port, 300);
+    }
+
+    #[test]
+    fn parse_port() {
+        let cmd = PoaNodeCommand::<()>::try_parse_from([
+            "reth",
+            "--discovery.port",
+            "300",
+            "--port",
+            "99",
+        ])
+        .unwrap();
+        assert_eq!(cmd.network.discovery.port, 300);
+        assert_eq!(cmd.network.port, 99);
+    }
+
+    #[test]
+    fn parse_metrics_port() {
+        let cmd = PoaNodeCommand::<()>::try_parse_from(["reth", "--metrics", "9001"]).unwrap();
+        assert_eq!(cmd.metrics, Some(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 9001)));
+
+        let cmd = PoaNodeCommand::<()>::try_parse_from(["reth", "--metrics", ":9001"]).unwrap();
+        assert_eq!(cmd.metrics, Some(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 9001)));
+
+        let cmd =
+            PoaNodeCommand::<()>::try_parse_from(["reth", "--metrics", "localhost:9001"]).unwrap();
+        assert_eq!(cmd.metrics, Some(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 9001)));
+    }
+
+    #[test]
+    fn parse_config_path() {
+        let cmd =
+            PoaNodeCommand::<()>::try_parse_from(["reth", "--config", "my/path/to/reth.toml"])
+                .unwrap();
+        // always store reth.toml in the data dir, not the chain specific data dir
+        let data_dir = cmd.datadir.unwrap_or_chain_default(cmd.chain.chain);
+        let config_path = cmd.config.unwrap_or(data_dir.config_path());
+        assert_eq!(config_path, Path::new("my/path/to/reth.toml"));
+
+        let cmd = PoaNodeCommand::<()>::try_parse_from(["reth"]).unwrap();
+
+        // always store reth.toml in the data dir, not the chain specific data dir
+        let data_dir = cmd.datadir.unwrap_or_chain_default(cmd.chain.chain);
+        let config_path = cmd.config.clone().unwrap_or(data_dir.config_path());
+        let end = format!("reth/{}/reth.toml", SUPPORTED_CHAINS[0]);
+        assert!(config_path.ends_with(end), "{:?}", cmd.config);
+    }
+
+    #[test]
+    fn parse_db_path() {
+        let cmd = PoaNodeCommand::<()>::try_parse_from(["reth"]).unwrap();
+        let data_dir = cmd.datadir.unwrap_or_chain_default(cmd.chain.chain);
+        let db_path = data_dir.db_path();
+        let end = format!("reth/{}/db", SUPPORTED_CHAINS[0]);
+        assert!(db_path.ends_with(end), "{:?}", cmd.config);
+
+        let cmd =
+            PoaNodeCommand::<()>::try_parse_from(["reth", "--datadir", "my/custom/path"]).unwrap();
+        let data_dir = cmd.datadir.unwrap_or_chain_default(cmd.chain.chain);
+        let db_path = data_dir.db_path();
+        assert_eq!(db_path, Path::new("my/custom/path/db"));
+    }
+
+    #[test]
+    fn parse_instance() {
+        let mut cmd = PoaNodeCommand::<()>::parse_from(["reth"]);
+        cmd.rpc.adjust_instance_ports(cmd.instance);
+        cmd.network.port = DEFAULT_DISCOVERY_PORT + cmd.instance - 1;
+        // check rpc port numbers
+        assert_eq!(cmd.rpc.auth_port, 8551);
+        assert_eq!(cmd.rpc.http_port, 8545);
+        assert_eq!(cmd.rpc.ws_port, 8546);
+        // check network listening port number
+        assert_eq!(cmd.network.port, 30303);
+
+        let mut cmd = PoaNodeCommand::<()>::parse_from(["reth", "--instance", "2"]);
+        cmd.rpc.adjust_instance_ports(cmd.instance);
+        cmd.network.port = DEFAULT_DISCOVERY_PORT + cmd.instance - 1;
+        // check rpc port numbers
+        assert_eq!(cmd.rpc.auth_port, 8651);
+        assert_eq!(cmd.rpc.http_port, 8544);
+        assert_eq!(cmd.rpc.ws_port, 8548);
+        // check network listening port number
+        assert_eq!(cmd.network.port, 30304);
+
+        let mut cmd = PoaNodeCommand::<()>::parse_from(["reth", "--instance", "3"]);
+        cmd.rpc.adjust_instance_ports(cmd.instance);
+        cmd.network.port = DEFAULT_DISCOVERY_PORT + cmd.instance - 1;
+        // check rpc port numbers
+        assert_eq!(cmd.rpc.auth_port, 8751);
+        assert_eq!(cmd.rpc.http_port, 8543);
+        assert_eq!(cmd.rpc.ws_port, 8550);
+        // check network listening port number
+        assert_eq!(cmd.network.port, 30305);
+    }
+
+    #[test]
+    fn with_unused_ports_conflicts_with_instance() {
+        let err = PoaNodeCommand::<()>::try_parse_from([
+            "reth",
+            "--with-unused-ports",
+            "--instance",
+            "2",
+        ])
+        .unwrap_err();
+        assert_eq!(err.kind(), clap::error::ErrorKind::ArgumentConflict);
+    }
+
+    #[test]
+    fn with_unused_ports_check_zero() {
+        let mut cmd = PoaNodeCommand::<()>::parse_from(["reth"]);
+        cmd.rpc = cmd.rpc.with_unused_ports();
+        cmd.network = cmd.network.with_unused_ports();
+
+        // make sure the rpc ports are zero
+        assert_eq!(cmd.rpc.auth_port, 0);
+        assert_eq!(cmd.rpc.http_port, 0);
+        assert_eq!(cmd.rpc.ws_port, 0);
+
+        // make sure the network ports are zero
+        assert_eq!(cmd.network.port, 0);
+        assert_eq!(cmd.network.discovery.port, 0);
+
+        // make sure the ipc path is not the default
+        assert_ne!(cmd.rpc.ipcpath, String::from("/tmp/reth.ipc"));
+    }
+}
